RDKB-10330 : For RxPowerLevel , datatype to be changed to int
Reason for change: Device.MoCA.Interface.1.AssociatedDevice.1.RxPowerLevel datatype to be changed to int from ulong
Test Procedure:
Query Device.MoCA.Interface.1.AssociatedDevice.1.RxPowerLevel
It should give negative value also.
Risks: None

+++ source/TR-181/include/cosa_moca_apis.h
@@ -355,7 +355,7 @@ struct _COSA_DML_MOCA_ASSOC_DEVICE
     ULONG                           PHYTxRate;
     ULONG                           PHYRxRate;
     ULONG                           TxPowerControlReduction;
-    ULONG                           RxPowerLevel;
+    INT                             RxPowerLevel;
     ULONG                           TxBcastRate;
     ULONG                           RxBcastPowerLevel;
     ULONG                           TxPackets;
